+++ SwiftReflector/Demangling/Swift4NodeToTLDefinition.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 
 namespace SwiftReflector.Demangling {
 	public class Swift4NodeToTLDefinition {
