+++ SwiftReflector/IOUtils/TempDirectoryStreamProvider.cs
@@ -5,7 +5,7 @@ using System;
 using System.IO;
 using System.Collections.Generic;
 using System.Linq;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using ObjCRuntime;
 
 namespace SwiftReflector.IOUtils {
