+++ SwiftReflector/Inventory/PropertyInventory.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
