+++ SwiftReflector/Inventory/ProtocolContents.cs
@@ -4,7 +4,7 @@
 using System;
 using System.IO;
 using System.Collections.Generic;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using SwiftReflector.Demangling;
 using ObjCRuntime;
