+++ SwiftReflector/SwiftXmlReflection/EnumDeclaration.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using System.Xml.Linq;
 using ObjCRuntime;
