+++ SwiftReflector/SwiftXmlReflection/ParameterItem.cs
@@ -6,7 +6,7 @@ using SwiftReflector.Inventory;
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using SwiftReflector.IOUtils;
 using ObjCRuntime;
