+++ SwiftReflector/TypeMapping/SwiftTypeToSLType.cs
@@ -3,7 +3,7 @@
 
 using System;
 using Dynamo.SwiftLang;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using System.Collections.Generic;
 using System.Linq;
 using ObjCRuntime;
