+++ SwiftReflector/TypeMapping/TypeMapper.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Globalization;
-using SwiftReflector.Exceptions;
+using SwiftReflector.ExceptionTools;
 using SwiftReflector.Inventory;
 using SwiftReflector.SwiftXmlReflection;
 using System.IO;
